requirements.txt
configer==1.3.1
configparser==7.1.0
fake-bpy-module-latest==20241010
filelock==3.16.1
freetype-py==2.5.1
fsspec==2024.9.0
human-body-prior==0.8.5.0
imageio==2.36.0
Jinja2==3.1.4
lazy_loader==0.4
MarkupSafe==3.0.1
mpmath==1.3.0
networkx==3.4.1
numpy==2.1.2
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
protobuf==5.28.3
pyglet==2.0.18
PyOpenGL==3.1.0
pyrender==0.1.45
scikit-image==0.24.0
scipy==1.14.1
six==1.16.0
smplx==0.1.28
sympy==1.13.3
tensorboardX==2.6.2.2
tifffile==2024.9.20
torch==2.4.1
torchgeometry==0.1.2
transforms3d==0.4.2
trimesh==4.5.2
typing_extensions==4.12.2
